Validation and Sanitization completed using express-validator

+++ app/controllers/office.controller.js
@@ -5,14 +5,45 @@ const Op = db.Sequelize.Op;
 
 const ERROR_MESSAGE = 'Something went wrong.'
 
+const { body, param, validationResult } = require('express-validator/check');
+
+exports.validate = (method) => {
+  switch (method) {
+    case 'createOfficeValidation': {
+     	return [ 
+        body('towerId', "Tower Id required").exists().isInt().trim().escape(),
+        body('office_number', "Office number required").exists().trim().escape().custom(reqOfficeNumber=> {
+          return new Promise((resolve, reject) => {
+            Office.findOne({ where: { office_number: reqOfficeNumber } })
+            .then(isExist => {
+              if(isExist !== null){
+                reject(new Error('Office already exists.'))
+              }else{
+                resolve(true)
+              }
+            })
+          })
+        }),
+        body('rent').optional().trim().escape(),
+      ]
+    }
+    case 'checkParamsIdValidation': {
+        return [
+          param('id', "Id required").exists().isInt(),
+        ]
+    }
+  }
+}
 
 exports.create = (req, res) => {
   
-   if (!req.body.towerId || !req.body.office_number) {
-	    return res.status(400).send({
-	      message: "Content can not be empty!"
-	    });   
-  	}
+  try{
+
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      res.status(422).json({ errors: errors.array() });
+      return;
+    }
 
     const office = {
       towerId: req.body.towerId,
@@ -29,16 +60,24 @@ exports.create = (req, res) => {
             message: err.message || ERROR_MESSAGE
           });
       });
-    
-
-};
+  }catch(err){
+    return res.status(500).send({ message: err.message || ERROR_MESSAGE });
+  }
+}
 
 exports.findAll = (req, res) => {
-  
+  try{
   	//Filters
-    const office_number = req.query.office_number;
-    let condition = office_number ? { office_number: { [Op.like]: `%${office_number}%` }, rent: { [Op.like]: `%${name}%` } } : null;
-    
+   
+    let condition = {};
+    if(req.query.towerId){
+      condition.towerId = req.query.towerId
+    }
+    if(req.query.office_number){
+      const office_number = req.query.office_number;
+      condition.office_number = { [Op.like]: `%${office_number}%` }
+    }
+
     //Pagination
     const page  = req.query.page || 0;
     const size  = req.query.size || 10;
@@ -50,13 +89,9 @@ exports.findAll = (req, res) => {
     const sortingOrder = [[sortBy, sortOrder]];
     
     //Get specific fields, Pass comma separated fields
-    const getAttributes = req.query.attributes ? req.query.attributes.split(',') : ['id', 'office_number', 'rent', 'createdAt'];
+    const getAttributes = req.query.attributes ? req.query.attributes.split(',') : ['id', 'office_number', 'rent', 'towerId', 'createdAt'];
 
-    Office.findAndCountAll({ where: condition, attributes:getAttributes, order:sortingOrder, limit, offset, 
-    	include:[{
-        	model: db.tower
-    	}]
-    })
+    Office.findAndCountAll({ where: condition, attributes:getAttributes, order:sortingOrder, limit, offset,})
     .then(data => {
     	const paginateData = getPagingData(data, page, limit);
       	return res.send(paginateData);
@@ -67,10 +102,18 @@ exports.findAll = (req, res) => {
       });
     });
 
-};
+  }catch(err){
+    return res.status(500).send({ message: err.message || ERROR_MESSAGE });
+  }
+}
 
 exports.findOne = (req, res) => {
-  
+  try{
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      res.status(422).json({ errors: errors.array() });
+      return;
+    }
     const id = req.params.id;
     Office.findByPk(id)
     .then(data => {
@@ -81,9 +124,18 @@ exports.findOne = (req, res) => {
         message: err.message || ERROR_MESSAGE
       });
     });
-};
+  }catch(err){
+    return res.status(500).send({ message: err.message || ERROR_MESSAGE });
+  }
+}
 
 exports.update = (req, res) => {
+  try{
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      res.status(422).json({ errors: errors.array() });
+      return;
+    }
   	const id = req.params.id;
   	Office.update(req.body, {
     where: { id: id }
@@ -104,9 +156,18 @@ exports.update = (req, res) => {
         message: err.message || ERROR_MESSAGE
       });
     });
-};
+  }catch(err){
+    return res.status(500).send({ message: err.message || ERROR_MESSAGE });
+  }
+}
 
 exports.delete = (req, res) => {
+  try{
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      res.status(422).json({ errors: errors.array() });
+      return;
+    }
   	const id = req.params.id;
   	Office.destroy({
     where: { id: id }
@@ -127,14 +188,17 @@ exports.delete = (req, res) => {
         message: err.message || ERROR_MESSAGE
       });
     });
-};
+  }catch(err){
+    return res.status(500).send({ message: err.message || ERROR_MESSAGE });
+  }
+}
 
 const getPagination = (page, size) => {
   const limit = size ? +size : 3;
   const offset = page ? page * limit : 0;
 
   return { limit, offset };
-};
+}
 
 const getPagingData = (data, page, limit) => {
   const { count: totalItems, rows: offices } = data;
@@ -142,4 +206,4 @@ const getPagingData = (data, page, limit) => {
   const totalPages = Math.ceil(totalItems / limit);
 
   return { totalItems, offices, totalPages, currentPage };
-};
+}
